Further refactor ChapterSelect
