+++ src/features/game/scenes/chapterSelect/ChapterSelect.ts
@@ -1,17 +1,13 @@
-import { screenCenter, Constants } from 'src/features/game/commons/CommonConstants';
+import { screenCenter } from 'src/features/game/commons/CommonConstants';
 import { limitNumber } from 'src/features/game/utils/GameUtils';
 import { addLoadingScreen } from '../../utils/LoadingScreen';
-import { fadeOut } from '../../effects/FadeEffect';
-import { ChapterDetail, SampleChapters } from './SampleChapters';
+import { SampleChapters } from './SampleChapters';
+import { ChapterDetail } from './ChapterSelectTypes';
 import { chapterSelectAssets, chapterSelectBackground } from './ChapterSelectAssets';
-import {
-  defaultScrollSpeed,
-  maskRect,
-  imageDist,
-  imageRect,
-  blackTintAlpha,
-  chapterSelectStyle
-} from './ChapterSelectConstants';
+import { defaultScrollSpeed, maskRect, imageDist } from './ChapterSelectConstants';
+import { createChapter } from './ChapterSelectHelper';
+import GameLayerManager from '../../layer/GameLayerManager';
+import { Layer } from '../../layer/GameLayerTypes';
 
 export type AccountInfo = {
   accessToken: string;
@@ -23,12 +19,16 @@ class ChapterSelect extends Phaser.Scene {
   private scrollSpeed: number;
   private chapterDetails: ChapterDetail[];
   private accountInfo: AccountInfo | undefined;
+  private layerManager: GameLayerManager;
 
   constructor() {
     super('ChapterSelect');
 
+    this.chapterContainer = undefined;
     this.scrollSpeed = defaultScrollSpeed;
     this.chapterDetails = SampleChapters;
+    this.accountInfo = undefined;
+    this.layerManager = new GameLayerManager();
   }
 
   init(accountInfo: AccountInfo) {
@@ -37,45 +37,13 @@ class ChapterSelect extends Phaser.Scene {
 
   public preload() {
     this.preloadAssets();
+    this.layerManager.initialiseMainLayer(this);
     addLoadingScreen(this);
-  }
-
-  private preloadAssets() {
-    chapterSelectAssets.forEach(asset => this.load.image(asset.key, asset.path));
-    this.chapterDetails.forEach((chapter, index) => {
-      this.load.image(`chapterImage${index}`, chapter.url);
-    });
   }
 
   public create() {
     this.renderBackground();
-    const mask = this.createMask();
-    this.chapterContainer = this.createChapterContainer();
-    this.add.existing(this.chapterContainer);
-
-    this.chapterContainer.mask = new Phaser.Display.Masks.GeometryMask(this, mask);
-  }
-
-  private renderBackground() {
-    this.add.image(screenCenter.x, screenCenter.y, chapterSelectBackground.key);
-  }
-
-  private createMask() {
-    const graphics = this.add.graphics();
-    const mask = graphics
-      .fillRect(maskRect.x, maskRect.y, maskRect.width, maskRect.height)
-      .setPosition(screenCenter.x, screenCenter.y);
-    mask.alpha = 0;
-    return mask;
-  }
-
-  private createChapterContainer() {
-    const chapterContainer = new Phaser.GameObjects.Container(this, 0, 0);
-    chapterContainer.add(
-      this.chapterDetails.map((chapterDetail, index) => createChapter(this, chapterDetail, index))
-    );
-
-    return chapterContainer;
+    this.renderChapters();
   }
 
   public update() {
@@ -98,62 +66,54 @@ class ChapterSelect extends Phaser.Scene {
     this.load.start();
   }
 
+  private preloadAssets() {
+    chapterSelectAssets.forEach(asset => this.load.image(asset.key, asset.path));
+    this.chapterDetails.forEach((chapter, index) => {
+      this.load.image(`chapterImage${index}`, chapter.previewBgPath);
+    });
+  }
+
+  private renderBackground() {
+    const background = new Phaser.GameObjects.Image(
+      this,
+      screenCenter.x,
+      screenCenter.y,
+      chapterSelectBackground.key
+    );
+    this.layerManager.addToLayer(Layer.Background, background);
+  }
+
+  private renderChapters() {
+    const mask = this.createMask();
+    this.chapterContainer = this.createChapterContainer();
+    this.chapterContainer.mask = new Phaser.Display.Masks.GeometryMask(this, mask);
+
+    this.layerManager.addToLayer(Layer.UI, this.chapterContainer);
+  }
+
+  private createMask() {
+    const graphics = this.add.graphics();
+    const mask = graphics
+      .fillRect(maskRect.x, maskRect.y, maskRect.width, maskRect.height)
+      .setPosition(screenCenter.x, screenCenter.y);
+    mask.alpha = 0;
+    return mask;
+  }
+
+  private createChapterContainer() {
+    const chapterContainer = new Phaser.GameObjects.Container(this, 0, 0);
+    chapterContainer.add(
+      this.chapterDetails.map((chapterDetail, index) => createChapter(this, chapterDetail, index))
+    );
+    return chapterContainer;
+  }
+
   private callGameManager(key: string) {
     if (key[0] === '#') {
       const text = this.cache.text.get(key);
       this.scene.start('GameManager', { text, accountInfo: this.accountInfo });
     }
   }
-}
-
-function createChapter(scene: ChapterSelect, { title, fileName }: ChapterDetail, index: number) {
-  const [x, y] = getCoorByChapter(index);
-  const image = new Phaser.GameObjects.Image(scene, 0, 0, `chapterImage${index}`).setDisplaySize(
-    imageRect.width,
-    imageRect.height
-  );
-  const blackTint = new Phaser.GameObjects.Rectangle(
-    scene,
-    0,
-    0,
-    imageRect.width,
-    imageRect.height,
-    0
-  ).setAlpha(blackTintAlpha);
-
-  const chapterText = `Chapter ${index}\n${title}`;
-  const text = new Phaser.GameObjects.Text(scene, 0, 0, chapterText, chapterSelectStyle).setOrigin(
-    0.5
-  );
-  const container = new Phaser.GameObjects.Container(scene, x, y, [image, blackTint, text]);
-  container.setSize(imageRect.width, imageRect.height);
-  container.setInteractive({
-    useHandCursor: true
-  });
-
-  container.on('pointerover', () => {
-    scene.add.tween(fadeOut([blackTint], Constants.fadeDuration * 2));
-  });
-
-  container.on('pointerout', () => {
-    scene.add.tween({
-      alpha: blackTintAlpha,
-      targets: blackTint,
-      duration: Constants.fadeDuration * 2
-    });
-  });
-
-  container.on('pointerdown', () => {
-    scene.loadFile(fileName);
-  });
-
-  return container;
-}
-
-function getCoorByChapter(chapterNum: number) {
-  const x = screenCenter.x + imageDist * chapterNum;
-  const y = screenCenter.y;
-  return [x, y];
 }
 
 export default ChapterSelect;
