+++ src/features/game/scenes/chapterSelect/ChapterSelectTypes.ts
@@ -1,0 +1,5 @@
+export type ChapterDetail = {
+  title: string;
+  previewBgPath: string;
+  fileName: string;
+};
