+++ src/features/game/scenes/chapterSelect/SampleChapters.ts
@@ -1,23 +1,19 @@
-export const SampleChapters = [
+import { ChapterDetail } from './ChapterSelectTypes';
+
+export const SampleChapters: ChapterDetail[] = [
   {
     title: 'Spaceship Emergency',
-    url: 'https://source.unsplash.com/random/500x700',
+    previewBgPath: 'https://source.unsplash.com/random/500x700',
     fileName: '../assets/chapter0.txt'
   },
   {
     title: 'Alien Attack',
-    url: 'https://source.unsplash.com/random/500x700',
+    previewBgPath: 'https://source.unsplash.com/random/500x700',
     fileName: '../assets/chapter1.txt'
   },
   {
     title: 'Jedi Master',
-    url: 'https://source.unsplash.com/random/500x700',
+    previewBgPath: 'https://source.unsplash.com/random/500x700',
     fileName: '../assets/chapter2.txt'
   }
 ];
-
-export type ChapterDetail = {
-  title: string;
-  url: string;
-  fileName: string;
-};
